InteractionTracker/requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.6
django-countries==5.0
django-filter==1.1.0
django-mathfilters==0.4.0
django-secure==1.0.1
djangorestframework==3.7.1
futures==3.1.1
idna==2.6
IndicoIo==1.1.1
Markdown==2.6.9
mock==1.3.0
nltk==3.2.5
olefile==0.44
pbr==3.1.1
Pillow==4.3.0
PyAudio==0.2.11
pypiwin32==220
pyttsx3==2.6
pytz==2017.2
requests==2.18.4
six==1.11.0
SpeechRecognition==3.7.1
textblob==0.13.0
urllib3==1.22
xmltodict==0.11.0
